AaharSense AI/requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
firebase-admin>=6.2.0
